requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-core
psycopg2-binary
python-dotenv
pgvector
flask
